requirements.txt
google-auth>=2.26.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.106.0
PySimpleGUI>=5.0.0
pyscard>=2.0.0
pyyaml>=6.0.0
keyboard>=0.13.0
PyInstaller>=6.0.0
